action and reducer created for applicant profile form
